refatorando e arrumando tipos de dados

+++ src/app/pages/dictionary-registration/home.ts
@@ -4,6 +4,7 @@ import { RouterLink } from '@angular/router';
 import { ModalDictionary } from './components/modal-dictionary/modal-dictionary';
 import { BsModalRef, BsModalService, ModalModule } from 'ngx-bootstrap/modal';
 import { ModalExcluir } from './components/modal-excluir/modal-excluir';
+import { IDictionary } from '../../shared/interfaces/dictionary.interfaces';
 
 
 
@@ -14,7 +15,7 @@ import { ModalExcluir } from './components/modal-excluir/modal-excluir';
   styleUrl: './home.css'
 })
 export class Home {
-  dictionaries = signal<any[]>([]);
+  dictionaries = signal<IDictionary[]>([]);
 
   dicionaryService = inject(DictionaryService);
 
@@ -41,18 +42,20 @@ export class Home {
     });
   }
 
-  deleteDictionaries(codigo: number) {
-  this.bsModalRef = this.modalService.show(ModalExcluir, {
-    initialState: {
-      message: 'Deseja excluir o Dicionário?',
-      onConfirm: () => {
-        this.dicionaryService.deleteDictionary(codigo).subscribe(() => {
-          this.getDicionaries();
-        });
+  deleteDictionaries(codigo: string | null) {
+    if (!codigo) return;
+    
+    this.bsModalRef = this.modalService.show(ModalExcluir, {
+      initialState: {
+        message: 'Deseja excluir o Dicionário?',
+        onConfirm: () => {
+          this.dicionaryService.deleteDictionary(codigo).subscribe(() => {
+            this.getDicionaries();
+          });
+        }
       }
-    }
-  });
-}
+    });
+  }
 
   editDictionary(dicionario: any) {
     this.dicionaryService.putDictionary(dicionario.id, dicionario).subscribe(() => {
